+++ packages/plugin-frontend/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+
+- Error banner for data synchronization
+
 ## [1.1.0] - 2024-04-24
 
 ## [1.0.0] - 2024-04-12